requirements.txt
beautifulsoup4==4.5.1
bs4==0.0.1
Django==1.10
image==1.5.4
Pillow==3.3.1
pytesseract==0.1.6
redis==2.10.5
requests==2.11.1
web.py==0.38
